Refactor About and Header components; update profile images and styles
