+++ src/components/header/header.css
@@ -39,7 +39,7 @@ header h5 {
   content: "";
   display: block;
   height: 1px;
-  width: 5rem;
+  /* width: 5rem; */
   background-color: var(
     --color-primary
   ); /* This is the color of the line. Change as desired */
@@ -113,6 +113,11 @@ header h5 {
     justify-content: center;
     width: 16rem;
     margin-left: 2.5rem;
+    margin-bottom: 2rem;
+    border-top-right-radius: 5rem;
+    border-top-left-radius: 1rem;
+    border-bottom-left-radius: 5rem;
+    border-bottom-right-radius: 1rem;
   }
 }
